src/app/components/confirmation-dialog/confirmation-dialog.component.ts
import { Component, OnInit } from "@angular/core";
import { MatDialogRef } from "@angular/material";

@Component({
  selector: "app-confirmation-dialog",
  templateUrl: "./confirmation-dialog.component.html",
  styleUrls: ["./confirmation-dialog.component.scss"]
})
export class ConfirmationDialogComponent implements OnInit {
  constructor(public dialogRef: MatDialogRef<ConfirmationDialogComponent>) {}

  ngOnInit() {}

  confirm() {
    this.dialogRef.close("confirmed");
  }

  decline() {
    this.dialogRef.close();
  }
}
